remove functions from headers * Apps
